add plumbing for services api name transformation
The MythTV Services API has a number of annoying
characteristics, one of which is that the element
names returned from a get or list request are not
always the same names one must use for a
modification (update/delete/add).  Sometimes the
difference is capitialization, but other more
interesting differences exist.

We will standardize on using the names returned
from the get/list request for our codes, but that
means we need a way to transforms those names to
the ones needed for an add/update/delete.  This
new function provides that plumbing.