An IconURL is not an Icon for the channel services api
The channel services api returns a IconURL, which is not
the actual filename of the icon.  The add/update is
expected to be provided a raw filename as the Icon.  We
cannot just return the IconURL as an Icon without some
breakage.  But as we do not get the filename we can't
avoid breakage in any case before channel services api
version 1.9 so this may be the least bad breakage.